#167861:NullPointerException at org.netbeans.modules.java.source.usages.LuceneIndex.clear

+++ java.source/src/org/netbeans/modules/java/source/usages/LuceneIndex.java
@@ -907,13 +907,15 @@ class LuceneIndex extends Index {
         try {
             final String[] content = this.directory.list();
             boolean dirty = false;
-            for (String file : content) {
-                try {
-                    directory.deleteFile(file);
-                } catch (IOException e) {
-                    //Some temporary files
-                    if (directory.fileExists(file)) {
-                        dirty = true;
+            if (content != null) {
+                for (String file : content) {
+                    try {
+                        directory.deleteFile(file);
+                    } catch (IOException e) {
+                        //Some temporary files
+                        if (directory.fileExists(file)) {
+                            dirty = true;
+                        }
                     }
                 }
             }
